Add available_moves_helper function to tidy up code inside queen, rook and bishop
