+++ src/engine/pieces/bishop.ts
@@ -3,7 +3,7 @@ import Player from '../player';
 import Board from '../board';
 import Square from '../square';
 import GameSettings from '../gameSettings';
-import King from './king';
+import { getAvailableMovesInDirections } from './available_moves_helper';
 
 export default class Bishop extends Piece {
     public constructor(player: Player) {
@@ -15,8 +15,6 @@ export default class Bishop extends Piece {
         const current_row: number = current_square.row;
         const current_col: number = current_square.col;
 
-        const available_moves: Square[] = [];
-
         const board_size:number = GameSettings.BOARD_SIZE;
 
 
@@ -27,30 +25,14 @@ export default class Bishop extends Piece {
             {row_move: -1, col_move: 1} 
         ]
 
-        for (var direction of directions) {
-            let row = current_row + direction.row_move;
-            let col = current_col + direction.col_move;
-
-            while (this.isOnBoard(board_size, row, col)) {
-
-                var new_square: Square = new Square(row, col);
-
-                var piece: Piece | undefined = board.getPiece(new_square);
-
-                if (!!piece) {
-                    if (piece.player !== this.player && !(piece instanceof King)) {
-                        available_moves.push(new_square);
-                    }
-                    break;
-                }
-
-                available_moves.push(new_square);
-
-                row += direction.row_move;
-                col += direction.col_move;
-            }
-
-        }
+        const available_moves: Square[] = getAvailableMovesInDirections(
+            directions,
+            current_row,
+            current_col,
+            this,
+            board_size,
+            board,
+        )
 
         return available_moves;
     }
