+++ src/engine/pieces/queen.ts
@@ -5,6 +5,8 @@ import Square from '../square';
 import GameSettings from '../gameSettings';
 import King from './king';
 import { Direction } from '../direction';
+import { getAvailableMovesInDirections } from './available_moves_helper';
+
 
 export default class Queen extends Piece {
     public constructor(player: Player) {
@@ -15,8 +17,6 @@ export default class Queen extends Piece {
         const current_square: Square = board.findPiece(this);
         const current_row: number = current_square.row;
         const current_col: number = current_square.col;
-
-        const available_moves: Square[] = [];
 
         const board_size: number = GameSettings.BOARD_SIZE;
 
@@ -31,30 +31,14 @@ export default class Queen extends Piece {
             {row_move: -1, col_move: 1}
         ]
 
-        for (var direction of directions) {
-            let row = current_row + direction.row_move;
-            let col = current_col + direction.col_move;
-
-            while (this.isOnBoard(board_size, row, col)) {
-
-                var new_square: Square = new Square(row, col);
-
-                var piece: Piece | undefined = board.getPiece(new_square);
-
-                if (!!piece) {
-                    if (piece.player !== this.player && !(piece instanceof King)) {
-                        available_moves.push(new_square);
-                    }
-                    break;
-                }
-
-                available_moves.push(new_square);
-
-                row += direction.row_move;
-                col += direction.col_move;
-            }
-
-        }
+        const available_moves: Square[] = getAvailableMovesInDirections(
+            directions,
+            current_row,
+            current_col,
+            this,
+            board_size,
+            board,
+        )
 
         return available_moves;
     }
