+++ src/engine/pieces/rook.ts
@@ -4,6 +4,7 @@ import Board from '../board';
 import Square from '../square';
 import GameSettings from '../gameSettings';
 import King from './king';
+import { getAvailableMovesInDirections } from './available_moves_helper';
 
 export default class Rook extends Piece {
     public constructor(player: Player) {
@@ -15,8 +16,6 @@ export default class Rook extends Piece {
         const current_row: number = current_square.row;
         const current_col: number = current_square.col;
 
-        const available_moves: Square[] = [];
-
         const board_size: number = GameSettings.BOARD_SIZE;
 
         const directions = [
@@ -26,30 +25,14 @@ export default class Rook extends Piece {
             {row_move: 0, col_move: -1}
         ]
 
-        for (var direction of directions){
-            let row = current_row + direction.row_move;
-            let col = current_col + direction.col_move;
-
-            while (this.isOnBoard(board_size, row, col)) {
-
-                var new_square: Square = new Square(row, col);
-
-                var piece: Piece | undefined = board.getPiece(new_square);
-
-                if (!!piece) {
-                    if (piece.player !== this.player && !(piece instanceof King)) {
-                        available_moves.push(new_square);
-                    }
-                    break;
-                }
-
-                available_moves.push(new_square);
-
-                row += direction.row_move;
-                col += direction.col_move;
-            }
-
-        }
+        const available_moves: Square[] = getAvailableMovesInDirections(
+            directions,
+            current_row,
+            current_col,
+            this,
+            board_size,
+            board,
+        )
 
         return available_moves;
     }
